[Feat/#4]: 게시글 조회 컨트롤러&서비스 핸들러 분리
